Move definition of TYPES to constants.
